% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_development.R
\name{egg_development}
\alias{egg_development}
\title{Egg development rate  as a function of temperature
.}
\usage{
egg_development(t)
}
\arguments{
\item{t}{Average daily temperature.}
}
\value{
Development rate (per day).
}
\description{
Egg development rate  as a function of temperature
.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
