% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_line.R
\name{fit_line}
\alias{fit_line}
\title{Fit a line of NDVI values between two dates.}
\usage{
fit_line(date1, date2, val1, val2)
}
\arguments{
\item{date1}{Date 1.}

\item{date2}{Date 2.}

\item{val1}{NDVI value at date1.}

\item{val2}{NDVI value at date2.}
}
\value{
Smoothed values.
}
\description{
Fit a line of NDVI values between two dates.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
