% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_land_cover.R
\name{get_land_cover}
\alias{get_land_cover}
\title{Get land cover type at location (lng, lat)}
\usage{
get_land_cover(lng, lat)
}
\arguments{
\item{lng}{Longitude of the coordinate of interest.}

\item{lat}{Latitude of the coordinate of interest.}
}
\value{
Land cover type id.
}
\description{
Get land cover type at location (lng, lat)
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
