% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NDVI_dnst.R
\name{get_NDVI_dnst}
\alias{get_NDVI_dnst}
\title{Get vegetation density type.}
\usage{
get_NDVI_dnst(x)
}
\arguments{
\item{xx}{NDVI value.}
}
\value{
Density type (Lowest, Lower, Dense, Higher or Highest).
}
\description{
Get vegetation density type.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
