% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NDVI_trend.R
\name{get_NDVI_trend}
\alias{get_NDVI_trend}
\title{Get NDVI trend.}
\usage{
get_NDVI_trend(xx, npsi = 10)
}
\arguments{
\item{xx}{NDVI dates and values.}

\item{npsi}{Number of change points in NDVI profile.}

\item{ndvi}{Vector of NDVI values.}
}
\value{
Trend, change date and values at the change pint and at the day.
}
\description{
Get NDVI trend.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
