% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NDVI_values.R
\name{get_NDVI_values}
\alias{get_NDVI_values}
\title{Get NDVI values between day1 and day2 at location (lng, lat)}
\usage{
get_NDVI_values(lng, lat, day1, day2)
}
\arguments{
\item{lng}{Longitude of the coordinate of interest.}

\item{lat}{Latitude of the coordinate of interest.}

\item{day1}{Start day of the period.}

\item{day2}{End day of the period.}
}
\value{
Vector with dates and NDVI values.
}
\description{
Get NDVI values between day1 and day2 at location (lng, lat)
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
