% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wind.R
\name{get_wind}
\alias{get_wind}
\title{Sample individual wind trajectory at day date and location (lng, lat)
.}
\usage{
get_wind(date, lng, lat, rn.sd)
}
\arguments{
\item{date}{Date hoppers hatched from eggs.}

\item{lng}{Longitude.}

\item{lat}{Latitude.}

\item{rn.sd}{Set random seed.}
}
\value{
Development rate (per day).
}
\description{
Sample individual wind trajectory at day date and location (lng, lat)
.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
