% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopper_devt.R
\name{hopper_dev}
\alias{hopper_dev}
\title{Hopper development conditions at location (lng, lat)
.}
\usage{
hopper_dev(date, lng, lat, thresh, peak = TRUE, lambda = 1, m = 2)
}
\arguments{
\item{date}{Date hoppers hatched from eggs.}

\item{lng}{Longitude.}

\item{lat}{Latitude.}

\item{thresh}{Minimal NDVI required for hopper development.}

\item{peak}{If account for appearance of fresh vegetation (default is TRUE).}

\item{lambda}{Parameter for smoothing NDVI.}

\item{m}{Number of change points in NDVI profile.}
}
\value{
Development rate (per day).
}
\description{
Hopper development conditions at location (lng, lat)
.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
