% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dev_period.R
\name{sample_dev_period}
\alias{sample_dev_period}
\title{Sample hopper development period from the truncated lognormal distribution}
\usage{
sample_dev_period(meanlog = log(36), sdlog = 1)
}
\arguments{
\item{meanlog}{Mean length of period (default is 36 days).}

\item{sdlog}{Standart deviation of period (default is q days).}
}
\value{
Length of egg development period in days
}
\description{
Sample hopper development period from the truncated lognormal distribution
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
