% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_Whittaker.R
\name{smooth_Whittaker}
\alias{smooth_Whittaker}
\title{Whittaker smoothing for NDVI values}
\usage{
smooth_Whittaker(y, lambda = 10^1.3)
}
\arguments{
\item{y}{Vector of NDVI values.}

\item{lambda}{Smoothing window.}
}
\value{
Smoothed values.
}
\description{
Whittaker smoothing for NDVI values
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
