% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suit_breeding.R
\name{suit_breeding}
\alias{suit_breeding}
\title{Calcualte suitability for breeding/laying eggs}
\usage{
suit_breeding(date, lng, lat, sbr, METdata)
}
\arguments{
\item{date}{Date}

\item{lng}{Longitude}

\item{lat}{Latitude}

\item{sbr}{Raster with breeding suitability scores}

\item{METdata}{MET Office data on soil moisture and precipitation}
}
\value{
Values of breeding suitability, total precipitation and max soil moisture
}
\description{
Calcualte suitability for breeding/laying eggs
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
