% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swrm_stay.R
\name{swrm_stay}
\alias{swrm_stay}
\title{Length of time a swarm stays at a location.}
\usage{
swrm_stay(lct, ndvi, npsi)
}
\arguments{
\item{lct}{Land cover type.}

\item{ndvi}{Profile of NDVI.}

\item{npsi}{Number of change points.}
}
\value{
Number of days.
}
\description{
Length of time a swarm stays at a location.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
