% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swrm_stay_type.R
\name{swrm_stay_type}
\alias{swrm_stay_type}
\title{Get type of stay based on land cover type and NDVI profile.}
\usage{
swrm_stay_type(lct, ndvi, npsi = 8)
}
\arguments{
\item{lct}{Land cover type.}

\item{ndvi}{Vector of NDVI values.}

\item{npsi}{Number of change points in NDVI profile.}
}
\value{
Type of stay: short, medium or long.
}
\description{
Get type of stay based on land cover type and NDVI profile.
}
\author{
Renata Retkute, \email{r.retkute@yahoo.com}
}
